==> notebooks/TIRE_dendritic_mouse/1A_generateSCE.Rmd <==
---
title: "Dendritic cell generate SCE object"
author: "Daniel Brown"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_depth: 3
    toc_float:
      smooth_scroll: true
    theme: readable
    highlight: tango 
    df_print: paged
    code_folding: hide
editor_options: 
    chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      dev = "png",
                      dpi = 300,
                      fig.asp=0.6, 
                      fig.path='./figures/generateSCE/',
                      warning=FALSE, message=FALSE)
knitr::opts_knit$set(root.dir = here::here())

library(tidyverse)
library(scuttle)
library(scran)
library(scater)
library(here)
library(edgeR)
library(biomaRt)
library(patchwork)
source(here::here("scripts/R/theme_publication.R"))
theme_set(theme_Publication())
```

# TurboCapture-Seq Dendritic cells Bcor KO + Flt3

## Recap

Prior 96w evaluation of TurboCapture-Seq v2 showed low UMIs recovered and high seq saturation. I did low throughput troubleshooting and didn’t see any issues. Process this experiment myself taking care to remove residual liquids from wash steps.

Process Hui Shi of Naik lab Bcor + Flt3 timecourse. Includes a few off my samples.

Samples

* Sorted dendritic cells
* HEK293T cell lysates in 1x TCL @ cells/uL
* PBMC cell lysates in 1x TCL @ 500 cells/uL
* No template control 1x TCL

# Generate count matrices for downstream analysis {.tabset}

## Mouse mapped reads from zUMI

Read in the sample sheets and metadata

```{r}
gene_names <- read_tsv(here::here(
  "data/GenomicsRnD/S000514/zUMI_outputs/expression/TIRE_DCs_Mm.gene_names.txt"))

meta_data <- read_csv(here::here(
  "metadata/S000442/DCs_Bcor_metadata_v1.1_HS.csv"))

mapping <- read_tsv(here::here(
  "data/GenomicsRnD/S000514/zUMI_outputs/stats/TIRE_DCs_Mm.readspercell.txt"
))
```

The number of reads is `r sum(mapping$N)`

Read the zUMI output

```{r read_zUMI}
zUMI_mixed <- readRDS(here::here
                ("data/GenomicsRnD/S000514/zUMI_outputs/expression/TIRE_DCs_Mm.dgecounts.rds"))

ds_names <- names(zUMI_mixed[["umicount"]][["exon"]][["downsampling"]])

# Check the number of samples for each downsampling
for (n in ds_names) {
  print(n)
  print(dim(zUMI_mixed[["umicount"]][["exon"]][["downsampling"]][[n]]))
}

counts <- zUMI_mixed[["umicount"]][["exon"]][["all"]]
dim(counts)
```

### Attach metadata

All samples are represented in the count matrix and metadata.  
In the pipeline I used 8bp of the well barcode sequence
The Amount column is in thousands of cells per uL so correct this.

```{r}
meta_data %>% 
  dplyr::select(Well_BC, Cell_type, Ligand) %>% 
  head()

```

Attach mapping rate metadata

```{r mapping_rate}
mapping <- mapping %>% 
  group_by(RG) %>% 
  pivot_wider(names_from = type, values_from = N) %>% 
  mutate(
    Reads = rowSums(across(where(is.numeric))),
    Mapped = 100 - (Unmapped / Reads * 100),
    Exon_Map = (Exon / Reads) * 100
    )

meta_data <- left_join(meta_data, mapping,
                by = c("Well_BC" = "RG"))
```

Subset the sample barcodes for those used in the experiment.

```{r}
keep_samples <- intersect(colnames(counts), meta_data$Well_BC)
discard_samples <- setdiff(colnames(counts), meta_data$Well_BC)

dropped <- meta_data[!meta_data$Well_BC %in% keep_samples,]
counts <- counts[,keep_samples]
meta_data <- meta_data[meta_data$Well_BC %in% keep_samples,]

meta_data <- meta_data %>% 
  arrange(Well_BC)
```

Check the sample order is the same for counts and metadata

```{r}
tail(colnames(counts))
tail(meta_data$Well_BC)
```

Check the discarded samples from the metadata. There are none.

```{r}
dropped
```

### Convert to SCE

```{r}
sce <- SingleCellExperiment(assays=list(counts=counts),
       colData = meta_data)
```

Add gene symbol to mouse

```{r}
gene_names <- gene_names %>% 
  arrange(gene_id)

gene_id <- as_tibble(row.names(sce))
colnames(gene_id) <- "gene_id"

geneID <- left_join(gene_id, gene_names,
                     by=join_by(gene_id))

rowData(sce)$ID <- geneID$gene_id
rowData(sce)$Symbol <- geneID$gene_name
row.names(sce) <- geneID$gene_name
```

### Compute QC metrics
Add per cell QC adds the QC info to the colData of the sce.

```{r}
is.mito.alt <- grepl("^mt", row.names(sce))

sce <- addPerCellQC(sce, subsets=list(
  Mito=is.mito.alt))
sce <- addPerFeatureQC(sce)
```

### Save SCE with quality control

```{r}
saveRDS(sce, here::here(
  "data/TIRE_dendritic_mouse/SCEs/", "DCs_basic.sce.rds"))

tb <- as_tibble(colData(sce))

write_csv(tb, here::here(
  "data/TIRE_dendritic_mouse/Matrices/", "DCs_coldata.csv"))

dge <- convertTo(
  sce, type ="edgeR")

saveRDS(dge, here::here(
  "data/TIRE_dendritic_mouse/SCEs/", "DCs_basic.dge.rds"))
```

### Session info
```{r}
sessionInfo()
```
