==> notebooks/TIRE_dendritic_mouse/1B_qualityControl_Vis.Rmd <==
---
title: "Dendritic cell visualise quality control"
author: "Daniel Brown"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_depth: 3
    toc_float:
      smooth_scroll: true
    theme: readable
    highlight: tango 
    df_print: paged
    code_folding: hide
editor_options: 
    chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      dev = "png",
                      dpi = 300,
                      fig.asp=0.6, 
                      fig.path='./figures/qc_metrics/',
                      warning=FALSE, message=FALSE)
knitr::opts_knit$set(root.dir = here::here())

library(SingleCellExperiment)
library(tidyverse)
library(scuttle)
library(scater)
library(platetools)
library(patchwork)
library(here)
source(here::here("scripts/R/theme_publication.R"))
theme_set(theme_Publication())
```

# TIRE-seq mouse dendritic cells

## Recap

Prior 96w evaluation of TurboCapture-Seq v2 showed low UMIs recovered and high seq saturation. I did low throughput troubleshooting and didn’t see any issues. Process this experiment myself taking care to remove residual liquids from wash steps.

Process Hui Shi of Naik lab Flt3 stimulated DC timecourse. Includes a few of my samples.

Samples

* Sorted dendritic cells
* HEK293T cell lysates in 1x TCL @ cells/uL
* PBMC cell lysates in 1x TCL @ 500 cells/uL
* No template control 1x TCL

## Notebook recap

SCE object in generate in 1A_generateSCE notebook.

## Notebook aim

Plot the QC metrics.

## Read SCE and extract col data


```{r}
sce <- readRDS(here::here(
   "data/TIRE_dendritic_mouse/SCEs/", "DCs_basic.sce.rds"))

tb <- as_tibble(colData(sce))

tb$Timepoint_Day <- as.factor(tb$Timepoint_Day)
```

# Check on plate map representation

Plot on plate map
Orientation is correct as low counts are where the no template control should be

```{r plate_sample, fig.width=12,fig.height=6}

p1 <- platetools::raw_map(data=tb$Cell_number, 
                     well=tb$Well, 
                     plate=96) + ggtitle("Cell input number")

p2 <- platetools::raw_map(data=log(tb$sum+1), 
                     well=tb$Well, 
                     plate=96) + ggtitle("Log library size")

p1 / p2
```

Remove NA samples

```{r}
tb <- tb[!is.na(tb$Cell_type),]
```

# High expressing genes

## High expressors

Well under 1 million counts which is the UMI n = 10 length. We are within the dynamic range of the assay.

```{r high_xpr}
plotHighestExprs(sce, n=10, as_percentage=F, colour_cells_by = "Ligand") + theme_Publication()
```

Look at the individual genes

```{r express}
plotExpression(sce, features = c("Gm10076", "Tmsb10", "Rps3", "Fau"), colour_by = "Ligand", exprs_values = "counts") + 
  theme_Publication() + scale_colour_Publication()
```

# Library size and gene plots {.tabset}

## Library size log

The text represents the timepoint. No clear pattern to me except maybe the genotype.

```{r libSize_log, fig.cap="Library size in deduplicated UMIs"}
plt1 <- ggplot(tb,
             aes(x = Cell_type, y= sum, colour = Genotype, label = Timepoint_Day, shape=Ligand)) + 
  geom_text() + 
  ylab("Library Size (UMIs)") + 
  xlab("Lysis buffer") +
  scale_y_continuous(trans='log10') +
  annotation_logticks(base = 10, sides = "l") +
  scale_fill_brewer(palette = "Dark2")

plt1
```

## Library size natural scale

```{r libSize_nat, fig.cap="Library size in deduplicated UMIs"}
plt2 <- ggplot(tb,
             aes(x = Cell_type, y= sum, colour = Genotype, label = Timepoint_Day, shape=Ligand)) +
  geom_text() + 
  ylab("Library Size (UMIs)") + 
  xlab("Lysis buffer") +  
  scale_fill_brewer(palette = "Dark2")

plt2
```

## Detected genes

```{r detected_genes_p, fig.cap="Number of detected genes"}
plt3 <- ggplot(tb,
             aes(x = Cell_type, y= detected, colour = Genotype, label = Timepoint_Day, shape=Ligand)) +
  geom_text() + 
  ylab("Genes detected") + 
  xlab("Lysis buffer") +
  scale_fill_brewer(palette = "Dark2")

plt3
```

## Genes v UMIs

Not much pattern here except HEKs and PBMCs from me. Fewer genes detected across comparable UMIs.

```{r gene_umi, fig.cap="UMIs versus genes"}
plt4 <- ggplot(tb, 
             aes(x = detected+1, y= sum+1, colour = Cell_type, label=Well)) + 
  geom_text() + 
  ylab("Library size (UMIs)") + 
  xlab("Genes detected") +
  scale_y_continuous(trans='log10') + scale_x_continuous(trans='log10') +
  annotation_logticks(base = 10, sides = "bl")

plt4
```

## Mitochondrial genes

Very low percentage for everything. Cell health looks good.

```{r mito_p, fig.cap="Percent mouse mitochondrial counts in library"}
plt5 <- ggplot(tb,
             aes(x = Cell_type, y= subsets_Mito_percent, colour = Genotype, label = Timepoint_Day, shape=Ligand)) +
  geom_text() + 
  ylab("Percent mouse mitochondrial genes") + 
  xlab("Lysis buffer") +
  scale_fill_brewer(palette = "Dark2")

plt5
```

# Conclusion

Quality of the sample looks good to me. Will further check by clustering samples
 
### Session info
```{r}
sessionInfo()
```
