==> notebooks/TIRE_dendritic_mouse/2B_Clustering.Rmd <==
---
title: "Dendritic cell sample clustering"
description: "wt cells only"
author: "Daniel Brown"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_depth: 3
    toc_float:
      smooth_scroll: true
    theme: readable
    highlight: tango 
    df_print: paged
    code_folding: hide
editor_options: 
    chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      dev = "png",
                      dpi = 300,
                      fig.asp=0.6, 
                      fig.path='./figures/clustering/',
                      warning=FALSE, message=FALSE)
knitr::opts_knit$set(root.dir = here::here())

library(SingleCellExperiment)
library(tidyverse)
library(scuttle)
library(scran)
library(scater)
library(edgeR)
library(patchwork)
library(knitr)
library(here)
source(here::here("scripts/R/theme_publication.R"))
theme_set(theme_Publication())
```

# TIRE-seq Dendritic cells Bcor Flt3 stimulation

## Recap

Prior 96w evaluation of TurboCapture-Seq v2 showed low UMIs recovered and high seq saturation. I did low throughput troubleshooting and didn’t see any issues. Process this experiment myself taking care to remove residual liquids from wash steps.

Process Hui Shi of Naik lab Bcor + Flt3 timecourse. Includes a few off my samples.

Samples

* Sorted dendritic cells
* HEK293T cell lysates in 1x TCL @ cells/uL
* PBMC cell lysates in 1x TCL @ 500 cells/uL
* No template control 1x TCL

## Notebook recap

SCE object in generate in 1A_generateSCE_reads notebook.

## Notebook aim

Check the samples cluster by their cell type. Only wild type cells to be clustered.

## Read SCE and preprocessing

This was generated in notebook 1A.

# Clustering

## All samples

```{r}
sce <- readRDS(here::here(
   "data/TIRE_dendritic_mouse/SCEs/DCs_basic.sce.rds"))
```

## Remove irrelatant samples

Remove my samples and KO samples as requested by Shalin

```{r}
sce_orig <- sce

sce <- sce[,!is.na(sce$Researcher)]
sce <- sce[,sce$Researcher == "HS"]
sce <- sce[,sce$Genotype == "WT"]
```

# Generate the principal component plot {.tabset}

## Cell type

```{r cell_type}
set.seed(666)
sce <- computeSumFactors(sce)
sce <- logNormCounts(sce)
sce <- runPCA(sce, ncomponents=2)

plt1 <- scater::plotPCA(sce,colour_by="Cell_type") + 
  theme_Publication(base_size=16)

plt1
```

## Timepoint

```{r time_point}
sce$Timepoint_Day <- as.factor(sce$Timepoint_Day)

plt2 <- scater::plotPCA(sce,colour_by="Timepoint_Day") + 
  theme_Publication(base_size=16)

plt2
```

## Combined PCAs

```{r combined, fig.asp=1.2}
plt1 / plt2
```

## Save data

```{r}
saveRDS(sce, here::here(
   "data/TIRE_dendritic_mouse/SCEs", "DCs_cluster.sce.rds"))
```

# Conclusion

The samples cluster like expected.

## Next steps

Differential expression testing
 
### Session info
```{r}
sessionInfo()
```
